Allow evolution chain to show pokemon alternates
Some pokemon species have multiple varieties. If these happen
in an evolution chain, the display of the pictures can show
all of them in a series if the user clicks through them. Each
click will advance to a new variety until the last which will
cycle back to the first one.

closes #1